requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
psycopg2==2.9.10
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
